% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.VernaBotaSims.R
\name{autoplot.VernaBotaSims}
\alias{autoplot.VernaBotaSims}
\title{plot the accuracy of scenarios}
\usage{
\method{autoplot}{VernaBotaSims}(object, ...)
}
\arguments{
\item{object}{an object of class ""VernaBotaSims"}

\item{...}{other arguments passed to or from other methods}
}
\value{
a ggplot of the accuracy of the different scenarios
}
\description{
plot the accuracy of scenarios
}
