% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSample.R
\name{CompareSample}
\alias{CompareSample}
\title{Perform simulation and compare different ways of sampling the data to fill}
\usage{
CompareSample(
  NbSamples = 3,
  Param = NULL,
  priors = NULL,
  D2fill,
  DAsso = NULL,
  pc2fill = NULL,
  pcFamilyDet = NULL,
  pcGenusDet = NULL,
  NbSim = 1,
  Results_Simulations = FALSE
)
}
\arguments{
\item{NbSamples}{an integer: the number of test dataset sampling to compare. All other parameters are the same parameters as for the function CompareSim.}

\item{Param}{a datatable with the parameters for each scenario :
\itemize{
\item priors : a vector with the rank of the priors to use in the priors list
\item dataAsso : a vector with the rank of the observation data to use in
the DAsso list (if Dasso is not provided, put 1)
\item weights : a vector with the weights of the priors
\item eps : a vector with the epsilon value for each scenario
\item Determ : a vector with the value of Determ (boolean)
}}

\item{priors}{a list of datasets containing the priors for each scenario.
These datasets must have been prepared using the function PrepPrior.
(Default is NULL: no prior information is used).}

\item{D2fill}{a dataset to fill, this dataset must
have been prepared using the function PrepData.}

\item{DAsso}{a list of datasets of observation used for each scenario, these datasets must have been
prepared using the function PrepData (Default is NULL: the dataset Data2fill is used
to built the association matrix).}

\item{pc2fill}{the percentage of data (dataFill) to fill.}

\item{pcFamilyDet}{the percentage of data determined
at the family level (from the subset of dataFill to fill).
We recommend using a percentage equivalent than in the data to gapfill.}

\item{pcGenusDet}{the percentage of data determined at
the genus level (from the subset of dataFill to fill, the rest isn't determined at all).
We recommend using a percentage equivalent than in the data to gapfill.}

\item{NbSim}{the number of simulations.}

\item{Results_Simulations}{a boolean specifying if the
user wants to keep the results of the simulations}
}
\value{
This function returns a list of 2 objects:
\itemize{
\item a list of objects of the class VernaBotaSims, resulting from the simulations
\item a datasets with one line per simulation : accuracy, scenario, and sampled test dataset
}
}
\description{
This function performs several simulations like CompareSim, but with different samples of the dataset used as test.
}
\details{
This function performs the following steps \emph{NbSamples} times:
\itemize{
\item Get the data,
\item Split between train and test set and remove taxonomic information from the test set (see \emph{SampleTestDataset} function),
\item for each scenario: perform simulations (see \emph{SimFullCom} function), compare simulations with original taxonomic information, with the function
\emph{CompareTaxo}, create an object of the class VernaBotaSims
}
}
