% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareTaxo.R
\name{CompareTaxo}
\alias{CompareTaxo}
\title{Compare the original taxonomic information of a dataset with predicted one}
\usage{
CompareTaxo(list_res, test_taxo)
}
\arguments{
\item{list_res}{Results of SimFullCom function applied to the test dataset:
a list of datatables with simulated taxonomic information.}

\item{test_taxo}{The original taxonomic information for the test dataset.}
}
\value{
A double corresponding to the percentage of well simulated species.
}
\description{
This function is used in the CompareSim function.
It compares the original taxonomic information of a dataset with predicted
one and return the percentage of matching species.
}
\details{
This function compares the taxonomic information in the column
GenSpCor of the simulated dataset list_res with the original information
from the columns Genus and Species of the dataset test_taxo.
The number of good estimation is divided by the length of the test dataset
to return a percentage good simulations.
}
