% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAlpha.R
\name{CreateAlpha}
\alias{CreateAlpha}
\title{Create matrix of posterior alpha}
\usage{
CreateAlpha(DataAsso, prior, wp)
}
\arguments{
\item{DataAsso}{datatable of inventory data, formatted as as shown in the vignette}

\item{prior}{datatable with prior expert knowledge, formatted as as shown in the vignette}

\item{wp}{numeric value giving the weighting of the prior information}
}
\value{
This function returns a datatable with a row per full botanical names (GenSp).
The first columns are GenSp, Family, Genus and Species and the following colums are each of the vernacular name.
These columns are the vectors of alphav.
}
\description{
This function creates a datatable containing the matrix
of posterior alpha and taxonomic info, using a prior expert knowledge (optional)
and observation for trees with a confirmed identification or
a temporary identification (BotaCertainty 4 and 3 in Guyafor)
}
\details{
This function performs the following steps:
\itemize{
\item create matrix of lambdav: get lambdav, equal to 1/number of possible association
to the vernacular name v if the association is possible according to the prior and 0 if not
\item create matrix of fv using DataObs
\item keep only trees with a confirmed identification or a temporary
identification (BotaCertainty 4 and 3 in Guyafor)
\item keep one line per individual (to avoid giving more weigth in individuals present in multiple censuses)
\item get fv, the observed frequencies of association between the vernacular name v and each botanical name fv
\item sum the two matrix with a weigh of wp for the prior and (1-wp) for the observation
}
}
