% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichCat.R
\name{DirichCat}
\alias{DirichCat}
\title{Draw a full species name using a Categorical-Dirichlet scheme}
\usage{
DirichCat(alphaV, Names, Vern)
}
\arguments{
\item{alphaV}{vector alphaV obtained in the DrawBota function (posterior alpha
for the given vernacular name, with all botanical name of Alpha)}

\item{Names}{vector of all botanical name of Alpha}

\item{Vern}{vernacular name}
}
\value{
A botanical name obtained from a Dirichlet-Categorical association.
}
\description{
This function draws a full species name for a given tree within the vector alphaV
using a Categorical-Dirichlet scheme.
}
\details{
This function performs the following steps:
\itemize{
\item draw one vector in a Dirichlet distribution of shape parameters Vect
\item draw one value in a Categorical distribution of parameter p the vector
obtained from the Dirichlet (i.e draw one botanical name one time with
the probability being the vector obtained from the Dirichlet)
\item retrieve the botanical name corresponding to the non null value in
this vector
}
}
