% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawBota.R
\name{DrawBota}
\alias{DrawBota}
\title{Gives a botanical name to a tree with a vernacular nameT}
\usage{
DrawBota(Alpha, VernName2Sim, Family2Sim, Genus2Sim, eps = 0.01, Determ)
}
\arguments{
\item{Alpha}{datatable containing the matrix of posterior alpha between botanical and
vernacular names, obtained with the funcion CreateAlpha}

\item{VernName2Sim}{the vernacular name of the tree for which we want to get a full botanical name}

\item{Family2Sim}{the family name of the tree for which we want to get a full botanical name}

\item{Genus2Sim}{the genus name of the tree for which we want to get a full botanical name}

\item{eps}{epsilon: background noise for species not associated with a given vernacular name.
Default is 0.01.}

\item{Determ}{boolean: if TRUE the more likely botanical names are return when a
association vernacular-botanical is performed.
If FALSE, the botanical names are drawn using a categorical-Dirichlet association scheme.}
}
\value{
This function returns a list with two elements: the first one is the associated botanical
name and the second one the method of association.
}
\description{
This function gives a botanical name to a tree with a vernacular name,
using a Dirichlet-Categorical method, and restricting the possible association according
to the family or the genus when these ahve been identified.
This function is performed within the function Get1Sim, for each tree without botanical
species name but with a vernacular name.
}
\details{
This function performs the following steps:
\itemize{
\item get alphaV a vector of posterior alpha for the given vernacular name, with all botanical name
of Alpha
\item give a null value in alphaV for for botanical name associated with this vernacular name
but from a different genus (if the genus is known) or a different family (if the family is known)
\item the the resulting alphaV contains only 0, then attributed the correction Det2Genus or Det2Fam
\item if there are non-null values in alphav, replace the null values by epsilon
divided by the number of null values and attributes a corrected name using the function
DrawDeterm if Determ is set to TRUE or the function DirichCat if Determ is set to FALSE
}
}
