% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.VernaBotaSims.R
\name{plot.VernaBotaSims}
\alias{plot.VernaBotaSims}
\title{plot the accuracy of scenarios}
\usage{
\method{plot}{VernaBotaSims}(x, ...)
}
\arguments{
\item{x}{an object of class "VernaBotaSims"}

\item{...}{other arguments passed to or from other methods}
}
\description{
plot the accuracy of scenarios
}
