% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{PrepData}
\alias{PrepData}
\title{Data preparation}
\usage{
PrepData(Data)
}
\arguments{
\item{Data}{data.frame of inventory data, formatted as it is when obtained
using the function EcoFoG::Guyafor2df or from the online data platform of
Paracou.}
}
\value{
a data.frame (formatted as a data.table) ready to be used for the
association of vernacular and botanical name. The column GenSp has been
added, which contains the Genus and species of the original data set (all
trees for which BotaSource=Vern have a GenSp="Indet.-Indet.")
}
\description{
This function prepares a dataset to be used for the association
of vernacular and botanical name.
}
\details{
This function performs the following steps:
\itemize{
\item convert all vernacular names to lower case (for consistency with the prior)
\item remove botanical names when BotaSource=Vern (for the reason explained
in the description of the function SimFullCom)
\item remove subspecies and variety names
\item add a column GenSp with the full botanical name
}
}
