% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorAllFG_20220126.R
\docType{data}
\name{PriorAllFG_20220126}
\alias{PriorAllFG_20220126}
\title{PriorAllFG_20220126}
\format{
A data.frame with 1657 rows and 197 variables:
\describe{
\item{Family}{Family name (character)}
\item{Genus}{Genus name (character)}
\item{Species}{Species name (character)}
\item{PresentInGuyaFor}{Is the taxon present in Guyafor database? (bolean)}
... vernacular names (0 if no possible association, 1 if possible
association according to expert knowledge)
}
}
\usage{
data(PriorAllFG_20220126)
}
\description{
Prior files resulting from the work of Jean-Maurice MadKaud,
for which the botanical and vernacular names have been checked
and updated on the 26/01/2022.
These prior file correspond to the prior file called "Probabilités Guyane
(sans erreur de relevé, facteur zone)" in Madkaud's report.
}
\keyword{datasets}
