% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorParacouNew_20220126.R
\docType{data}
\name{PriorParacouNew_20220126}
\alias{PriorParacouNew_20220126}
\title{PriorParacouNew_20220126}
\format{
A data.frame with 661 rows and 196 variables:
\describe{
\item{Family}{Family name (character)}
\item{Genus}{Genus name (character)}
\item{Species}{Species name (character)}
\item{PresentInGuyaFor}{Is the taxon present in Guyafor database? (bolean)}
... vernacular names (0 if no possible association, 1 if possible
association according to expert knowledge)
}
}
\usage{
data(PriorParacouNew_20220126)
}
\description{
Prior files resulting from the work of Jean-Maurice MadKaud (see Vignette),
for which the botanical and vernacular names have been checked
and updated on the 26/01/2022.
This prior file correspond to the prior file called
"Probabilités Paracou (avec erreur de relevé, facteur temps (équipe))" in
Madkaud's report.
}
\keyword{datasets}
