% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleTestDataset.R
\name{SampleTestDataset}
\alias{SampleTestDataset}
\title{Split data between training and test}
\usage{
SampleTestDataset(dat, pc2fill, pcGenusDet, pcFamilyDet)
}
\arguments{
\item{dat}{The data to split}

\item{pc2fill}{the percentage of data (dat) to fill}

\item{pcGenusDet}{the percentage of data determined at
the genus level (from the subset of dataFill to fill, the rest isn't determined at all).}

\item{pcFamilyDet}{the percentage of data determined
at the family level (from the subset of dataFill to fill).}
}
\value{
This function returns a list with 3 elements:
\itemize{
\item The dataset with some taxonomis information removed
\item The taxonomic information that has been removed
\item The corresponding tree ID (trees from the test dataset)
}
}
\description{
This function is used in the \emph{CompareSim} and \emph{CompareSample} functions.
It splits the data between train an test datasets, and remove the taxonomic information from the test dataset.
}
\details{
This function
\itemize{
\item Split between train and test set according to the parameter pc2fill (keep only fully identified trees in the test set),
\item Remove taxonomic information from the test set (at the species, genus, or family level according to the parameters pcFamilyDet and pcGenusDet),
}

@export
}
