% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabletest.VernaBotaSims.R
\name{stabletest}
\alias{stabletest}
\title{test the stability of the results of CompareSim}
\usage{
stabletest(object, s, ...)
}
\arguments{
\item{object}{an object of class "VernaBotaSims"}

\item{s}{the scenario for which the user want to test the stability of the simulations}

\item{...}{other arguments passed to or from other methods}
}
\description{
test the stability of the results of CompareSim
}
