% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.VernaBotaSims.R
\name{summary}
\alias{summary}
\title{summary S3 method for class "VernaBotaSims"}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{an object of class ""VernaBotaSims"}

\item{...}{other arguments passed to or from other methods}
}
\value{
a summary of the scenarios tested
}
\description{
summary S3 method for class "VernaBotaSims"
}
