% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.VernaBotaSims.R
\name{plot.VernaBotaSims}
\alias{plot.VernaBotaSims}
\title{plot S4 method for class "VernaBotaSims"}
\usage{
\method{plot}{VernaBotaSims}(x, ...)
}
\arguments{
\item{x}{an object of class ""VernaBotaSims"}

\item{...}{other arguments passed to or from other methods}
}
\value{
a plot of the accuracy of the different scenarios
}
\description{
plot S4 method for class "VernaBotaSims"
}
