% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{PrepData}
\alias{PrepData}
\title{Data preparation}
\usage{
PrepData(Data)
}
\arguments{
\item{Data}{dataframe of inventory data, formatted as as shown in the vignette}
}
\value{
a dataframe (formatted as a data.table) ready to be used for the association of vernacular and botanical name
}
\description{
This function prepares a dataset to be used for the association of vernacular and botanical name.
}
\details{
This function performs the following steps:
\itemize{
\item convert all vernacular names to lower case (for consistency with the prior)
\item remove botanical names when BotaSource=Vern (for the reason explained
in the description of the function SimFullCom)
\item remove subspecies and variety names
\item add a column GenSp with the full botanical name
}
}
