% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.VernaBotaSims.R
\name{summary.VernaBotaSims}
\alias{summary.VernaBotaSims}
\title{summary S4 method for class "VernaBotaSims"}
\usage{
\method{summary}{VernaBotaSims}(object, ...)
}
\arguments{
\item{object}{an object of class ""VernaBotaSims"}

\item{...}{other arguments passed to or from other methods}
}
\value{
a summary of the scenarios tested
}
\description{
summary S4 method for class "VernaBotaSims"
}
