% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vernabota-package}
\alias{vernabota}
\alias{vernabota-package}
\title{vernabota: Association between vernacular and botanical names for Guyafor data}
\description{
This package allow gapfilling missing botanical names using vernacular names, in the case of Guyafor census data. The objective is to obtain a chosen number of simulated communities for which individuals only identified with a vernacular name are given a botanical name from association of vernacular and botanical names in a Dirichlet-categorical scheme.
}
\author{
\strong{Maintainer}: Geraldine Derroire \email{geraldine.derroire@cirad.fr}

Authors:
\itemize{
  \item Melaine Aubry-Kientz
  \item Ariane Mirabel
  \item Daniela Krebber
  \item Eric Marcon
  \item Bruno Herault
}

}
\keyword{internal}
