% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VernaBotaSims.R
\docType{class}
\name{VernaBotaSims-class}
\alias{VernaBotaSims-class}
\alias{VernaBotaSims}
\title{VernaBotaSims class and associated methods}
\description{
The VernaBotaSims class is the output of the CompareSim function.

Plot function to be applied on \emph{VernaBotaSims} objects. This function draw
a boxplot of the percentages of well associated species for each scenario.

Summary function to be applied on \emph{VernaBotaSims} objects. This function print
the results of each scenario, together with its parameters.
}
\section{Slots}{

\describe{
\item{\code{NScenar}}{numeric: number of scenario simulated}

\item{\code{ParamScenar}}{data.frame: a datatable with the parameters for each
scenario (input of CompareSim)}

\item{\code{D2fill}}{list of datasets to fill for each scenario (input of CompareSim)}

\item{\code{DAsso}}{list of datasets of observation used for each scenario (input of
CompareSim)}

\item{\code{priors}}{list of datasets containing the priors for each scenario (input
of CompareSim)}

\item{\code{pc_results}}{list of double corresponding to the percentage of well
simulated species}

\item{\code{results}}{list of results from SimFullCom for each scenario.}
}}

